+++ Loader/SwaggerPhpConfigurationLoader.php
@@ -14,7 +14,10 @@ declare(strict_types=1);
 namespace Linkin\Bundle\SwaggerResolverBundle\Loader;
 
 use EXSyst\Component\Swagger\Swagger;
+use Linkin\Bundle\SwaggerResolverBundle\Collection\SchemaDefinitionCollection;
 use Linkin\Bundle\SwaggerResolverBundle\Merger\OperationParameterMerger;
+use Swagger\Annotations\Swagger as SwaggerZircote;
+use Symfony\Component\Config\Resource\FileResource;
 use Symfony\Component\Routing\RouterInterface;
 
 use function json_decode;
@@ -36,6 +39,11 @@ class SwaggerPhpConfigurationLoader extends AbstractAnnotationConfigurationLoade
     private $scan;
 
     /**
+     * @var SwaggerZircote
+     */
+    private $swaggerAnnotation;
+
+    /**
      * @param OperationParameterMerger $merger
      * @param RouterInterface $router
      * @param array $scan
@@ -54,10 +62,20 @@ class SwaggerPhpConfigurationLoader extends AbstractAnnotationConfigurationLoade
      */
     protected function loadConfiguration(): Swagger
     {
-        $swaggerAnnotation = scan($this->scan, [
+        $this->swaggerAnnotation = scan($this->scan, [
             'exclude' => $this->exclude,
         ]);
 
-        return new Swagger(json_decode((string) $swaggerAnnotation, true));
+        return new Swagger(json_decode((string) $this->swaggerAnnotation, true));
+    }
+
+    protected function registerDefinitionResources(SchemaDefinitionCollection $definitionCollection): void
+    {
+        foreach ($this->swaggerAnnotation->definitions as $zircoteDefinition) {
+            $definitionCollection->addSchemaResource(
+                $zircoteDefinition->definition,
+                new FileResource($zircoteDefinition->_context->filename)
+            );
+        }
     }
 }