+++ Tests/Fixtures/FixturesProvider.php
@@ -28,14 +28,32 @@ class FixturesProvider
     ];
 
     private const MAP_ROUTE_RESOURCE = [
-        'customers_get' => __DIR__ . '/SwaggerPhp/Controllers/CustomerController.php',
-        'customers_post' => __DIR__ . '/SwaggerPhp/Controllers/CustomerController.php',
-        'customers_get_one' => __DIR__ . '/SwaggerPhp/Controllers/CustomerController.php',
-        'customers_update' => __DIR__ . '/SwaggerPhp/Controllers/CustomerController.php',
-        'customers_patch' => __DIR__ . '/SwaggerPhp/Controllers/CustomerController.php',
-        'customers_delete' => __DIR__ . '/SwaggerPhp/Controllers/CustomerController.php',
-        'customers_password_create' => __DIR__ . '/SwaggerPhp/Controllers/CustomerPasswordController.php',
-        'customers_password_reset' => __DIR__ . '/SwaggerPhp/Controllers/CustomerPasswordController.php',
+        'customers_get' => [
+            __DIR__ . '/SwaggerPhp/Controllers/CustomerController.php',
+        ],
+        'customers_post' => [
+            __DIR__ . '/SwaggerPhp/Models/CustomerNew.php',
+            __DIR__ . '/SwaggerPhp/Controllers/CustomerController.php',
+        ],
+        'customers_get_one' => [
+            __DIR__ . '/SwaggerPhp/Controllers/CustomerController.php',
+        ],
+        'customers_update' => [
+            __DIR__ . '/SwaggerPhp/Models/CustomerNew.php',
+            __DIR__ . '/SwaggerPhp/Controllers/CustomerController.php',
+        ],
+        'customers_patch' => [
+            __DIR__ . '/SwaggerPhp/Controllers/CustomerController.php',
+        ],
+        'customers_delete' => [
+            __DIR__ . '/SwaggerPhp/Controllers/CustomerController.php',
+        ],
+        'customers_password_create' => [
+            __DIR__ . '/SwaggerPhp/Controllers/CustomerPasswordController.php',
+            ],
+        'customers_password_reset' => [
+            __DIR__ . '/SwaggerPhp/Controllers/CustomerPasswordController.php',
+        ],
     ];
 
     private const MAP_PATH_TO_ROUTE = [
@@ -65,7 +83,7 @@ class FixturesProvider
         return self::MAP_PATH_TO_ROUTE[$path][$method];
     }
 
-    public static function getResourceByRouteName(string $routeName): string
+    public static function getResourceByRouteName(string $routeName): array
     {
         return self::MAP_ROUTE_RESOURCE[$routeName];
     }